refactor(fish, food): Improve performance (#12)
Improves performance by removing loops, and instead executing logic per
fish and food per frame. This fixes the CPU spiking issue, while also
allowing us to fit a lot more fish and food on screen at once. There are
also a few bug fixes in here, along with variable fish speed.

+++ src/coin.h
@@ -22,7 +22,7 @@ typedef struct
 } Coin;
 
 extern Coin coin_list[MAX_COINS];
-extern uint16_t player_coins;
+extern uint8_t player_coins;
 
 void
 init_coins(void);
